% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.ci.nbPearson.R
\name{bss.ci.nbPearson}
\alias{bss.ci.nbPearson}
\title{Bayesian sample size using ACC or ALC criterion for the negative binomial/Pearson Type VI model}
\usage{
bss.ci.nbPearson(crit, lam0, theta0, phi, w, rho, len = NULL,
  len.max = NULL, R = 1000, n0 = 1)
}
\arguments{
\item{crit}{A characther string specifying the criterion. Criteria: "ACC" and "ALC".}

\item{lam0}{A positive real number representing a shape parameter of the prior distribution.}

\item{theta0}{A positive real number representing a shape parameter of the prior distribution.}

\item{phi}{A positive real number representing a scale parameter of the prior distribution.}

\item{w}{A positive real number representing the aliquot volume.}

\item{rho}{A number in (0, 1). The probability of the credible interval is equal or greater than $1-rho$ depending on the criterion used.}

\item{len}{A positive real number representing the length of the credible intervals in the ACC criterion.}

\item{len.max}{A positive real number representing the maximum length of the credible  intervals in the ALC criterion.}

\item{R}{Number of replicates used in the simulation. Default is 1000.}

\item{n0}{A positive integer representing the initial sample size in which the function will check the criterion. Default is 1.}
}
\value{
An integer representing the sample size.
}
\description{
Bayesian sample size using ACC or ALC criterion for the negative binomial/Pearson Type VI model
}
