% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.ci.poidp.R
\name{bss.ci.PoiDP}
\alias{bss.ci.PoiDP}
\title{Bayesian sample size using ACC or ALC criterion for the Poisson/Dirichlet process model}
\usage{
bss.ci.PoiDP(crit, lam0, theta0, w, rho, alpha, len = NULL,
  len.max = NULL, eps = NULL, cgrid = 0.01, R = 100, n0 = 2,
  inc = c(100, 10, 5))
}
\arguments{
\item{crit}{A characther string specifying the criterion. Criteria: "ACC" and "ALC".}

\item{lam0}{A positive real number representing a hyperparameter of the $F_0$ base distribution.}

\item{theta0}{A positive real number representing a hyperparameter of the $F_0$ base distribution. We consider $F_0$ as the gamma distribution with mean $lam_0$ and shape parameter $theta_0$.}

\item{w}{A positive real number representing the aliquot volume.}

\item{rho}{A number in (0, 1). The probability of the credible interval is equal or
greater than $1-rho$ depending on the criterion used.}

\item{alpha}{Shape parameter of the Dirichlet process.}

\item{len}{A positive real number representing the length of the credible intervals in the
ACC criterion.}

\item{len.max}{A positive real number representing the maximum length of the credible 
intervals in the ALC criterion.}

\item{eps}{...}

\item{cgrid}{Length of the grid used in the points for which the probability is computed.}

\item{R}{Number of replicates used in the simulation. Default is 100.}

\item{n0}{A positive integer representing the initial sample size in which the function 
will check the criterion. Default is 1.}

\item{inc}{...}
}
\value{
An integer representing the sample size.
}
\description{
Bayesian sample size using ACC or ALC criterion for the Poisson/Dirichlet process model
}
