% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.dt.nbpearson.R
\name{bss.dt.nbPearson}
\alias{bss.dt.nbPearson}
\title{Bayesian sample size in a decision-theoretic approach for the negative binomial/Pearson Type VI model}
\usage{
bss.dt.nbPearson(lf, lam0, theta0, phi, w, c, rho = NULL, gam = NULL,
  nmax = 100, nrep = 10, lrep = 50, plot = TRUE, ...)
}
\arguments{
\item{lf}{1 or 2, representing the loss function used.}

\item{lam0}{A positive real number representing the prior expected value for the prior 
gamma distribution.}

\item{theta0}{A positive real number representing the shape parameter for the prior 
gamma distribution.}

\item{phi}{A positive real number representing a scale parameter of the prior distribution.}

\item{w}{A positive real number representing the aliquot volume.}

\item{c}{A positive real number representing the cost of colect one aliquot.}

\item{rho}{A number in (0, 1). The probability of the credible interval is $1-rho$. Only
for lost function 1.}

\item{gam}{A positive real number connected with the credible interval when using lost
function 2.}

\item{nmax}{A positive integer representing the maximum number for compute the Bayes risk.
Default is 100.}

\item{nrep}{A positive integer representing the number of samples taken for each $n$.}

\item{lrep}{A positive integer representing the number of samples taken for $S_n$.}

\item{plot}{Boolean. If TRUE (default) it plot the estimated Bayes risks and the fitted
curve.}

\item{...}{Currently ignored.}
}
\value{
An integer representing the sample size.
}
\description{
Bayesian sample size in a decision-theoretic approach for the negative binomial/Pearson Type VI model
}
