% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.dt.poidp.R
\name{bss.dt.PoiDP}
\alias{bss.dt.PoiDP}
\title{Bayesian sample size in a decision-theoretic approach for the Poisson/Dirichlet process.}
\usage{
bss.dt.PoiDP(lam0, theta0, alpha, w, c, nmax = 100, nrep = 10,
  R = 100, plot = TRUE, ...)
}
\arguments{
\item{lam0}{A positive real number representing a hyperparameter of the $F_0$ base distribution.}

\item{theta0}{A positive real number representing a hyperparameter of the $F_0$ base distribution. We consider $F_0$ as the gamma distribution with mean $lam_0$ and shape parameter $theta_0$.}

\item{alpha}{Shape parameter of the Dirichlet process.}

\item{w}{A positive real number representing the aliquot volume.}

\item{c}{A positive real number representing the cost of colect one aliquot.}

\item{nmax}{A positive integer representing the maximum number for compute the Bayes risk.
Default is 100.}

\item{nrep}{A positive integer representing the number of samples taken for each $n$.}

\item{R}{Number of replicates used in the simulation. Default is 100.}

\item{plot}{Boolean. If TRUE (default) it plot the estimated Bayes risks and the fitted
curve.}

\item{...}{Currently ignored.}
}
\value{
An integer representing the sample size.
}
\description{
Bayesian sample size in a decision-theoretic approach for the Poisson/Dirichlet process.
}
