% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlam.R
\name{rlam}
\alias{rlam}
\title{Random generation for the Dirichlet process with a gamma distribution as the $F_0$ base distribution.}
\usage{
rlam(n, alpha, lam0, theta0, eps = 0.01)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Shape parameter of the Dirichlet process.}

\item{lam0}{A positive real number representing a hyperparameter of the $F_0$ base distribution.}

\item{theta0}{A positive real number representing a hyperparameter of the $F_0$ base distribution. We consider $F_0$ as the gamma distribution with mean $lam_0$ and shape parameter $theta_0$.}

\item{eps}{Tolerance limit used in the simulation algorithm. Default is 0.01.}
}
\value{
A sample of size n from the Dirichlet process.
}
\description{
Random generation for the Dirichlet process with a gamma distribution as the $F_0$ base distribution.
}
