% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.ee.R
\name{ss.ee}
\alias{ss.ee}
\title{Sample size for controling the estimation error controling the maximum absolute error,
the maximum relative error or both}
\usage{
ss.ee(dist, rho, a, b, ea = NULL, er = NULL, w = NULL, phi = NULL,
  n0 = 2, lag = 1)
}
\arguments{
\item{dist}{A characther string specifying the probability distribution. Distributions: 
"binomial", "negative binomial" and "poisson"}

\item{rho}{A number in (0, 1). The probability 1-rho represents the minimum 
confidence level}

\item{a}{A positive real number representing the lower bound of the interval (a, b)}

\item{b}{A positive real number representing the upper bound of the interval (a, b)}

\item{ea}{A positive real number representing the maximum admissible absolute estimation 
error}

\item{er}{A positive real number representing the maximum admissible relative estimation 
error}

\item{w}{A positive real number representing the aliquot volume (only available for 
negative binomial distribution). Default is NULL}

\item{phi}{A positive real number representing the shape parameter for negative 
binomial distribution. Default is NULL}

\item{n0}{A positive integer representing the initial sample size in which the function 
will check the criterion. Default is 2.}

\item{lag}{A positive integer representing the increment which 'n' receives until reach
the criterion. Default is 1.}
}
\value{
An integer representing the sample size
}
\description{
Sample size for controling the estimation error controling the maximum absolute error,
the maximum relative error or both
}
\references{
Costa, E. G., Lopes, R. M. & Singer, J. M. (2016). Sample size for estimating 
the mean concentration of organisms in ballast water. Journal of Environmental Management 
180, 433–438.

Chen, X. (2011). Exact computation of minimum sample size for estimation of 
binomial parameters. Journal of Statistical Planning and Inference 141, 2622–2632.

Chen, Z. & Chen, X. (2016). Exact calculation of minimum sample size for 
estimating a Poisson parameter. Communications in Statistics-Theory and Methods 45, 
4692–4715
}
