% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlambar.R
\name{rlambar}
\alias{rlambar}
\title{Random generation for the functional mean of the Dirichlet process with a gamma distribution as the $F_0$ base distribution.}
\usage{
rlambar(N, alpha, lam0, theta0, eps = 0.1, ncore = NULL)
}
\arguments{
\item{N}{Number of observations.}

\item{alpha}{Shape parameter of the Dirichlet process.}

\item{lam0}{A positive real number representing a hyperparameter of the $F_0$ base distribution.}

\item{theta0}{A positive real number representing a hyperparameter of the $F_0$ base distribution. We consider $F_0$ as the gamma distribution with mean $lam_0$ and shape parameter $theta_0$.}

\item{eps}{Tolerance limit used in the simulation algorithm.}

\item{ncore}{Number of cores to use in parallel computin. If NULL the function uses 1 core if there is only one core, if there is more than one cores uses one half of the cores.}
}
\value{
A random sample of the functional mean of the Dirichlet process.
}
\description{
Random generation for the functional mean of the Dirichlet process with a gamma distribution as the $F_0$ base distribution.
}
