% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlambar.xn.R
\name{rlambar.xn}
\alias{rlambar.xn}
\title{Random generation for the posterior functional mean of the Poisson/Dirichlet process with a gamma distribution as the $F_0$ base distribution.}
\usage{
rlambar.xn(N, alpha, x, w, lam0, theta0)
}
\arguments{
\item{N}{Number of observations.}

\item{alpha}{Shape parameter of the Dirichlet process.}

\item{x}{Observed counts of the Poisson distribution of the model.}

\item{w}{A positive real number representing the aliquot volume.}

\item{lam0}{A positive real number representing a hyperparameter of the $F_0$ base distribution.}

\item{theta0}{A positive real number representing a hyperparameter of the $F_0$ base distribution. We consider $F_0$ as the gamma distribution with mean $lam_0$ and shape parameter $theta_0$.}
}
\value{
A random sample of the posterior functional mean of the Poisson/Dirichlet process.
}
\description{
Random generation for the posterior functional mean of the Poisson/Dirichlet process with a gamma distribution as the $F_0$ base distribution.
}
